Split the services package into two: oauth, and remoteservices.
